refactor(e2e-playground): extract consent-denied negative-path into shared social-login helper
